// client/src/redux/profile/profile.types.ts
export enum ProfileTypes {
  SEARCH_PROFILES_START = "SEARCH_PROFILES_START",
  SEARCH_PROFILES_SUCCESS = "SEARCH_PROFILES_SUCCESS",
  SEARCH_PROFILES_FAILURE = "SEARCH_PROFILES_FAILURE",
  FETCH_PROFILE_START = "FETCH_PROFILE_START",
  FETCH_PROFILE_SUCCESS = "FETCH_PROFILE_SUCCESS",
  FETCH_PROFILE_FAILURE = "FETCH_PROFILE_FAILURE",
  RESET_SELECTED_PROFILE = "RESET_SELECTED_PROFILE",
  FETCH_PROFILE_FRIENDS_START = "FETCH_PROFILE_FRIENDS_START",
  FETCH_PROFILE_FRIENDS_SUCCESS = "FETCH_PROFILE_FRIENDS_SUCCESS",
  FETCH_PROFILE_FRIENDS_FAILURE = "FETCH_PROFILE_FRIENDS_FAILURE",
  SEND_FRIEND_REQUEST_START = "SEND_FRIEND_REQUEST_START",
  SEND_FRIEND_REQUEST_SUCCESS = "SEND_FRIEND_REQUEST_SUCCESS",
  SEND_FRIEND_REQUEST_FAILURE = "SEND_FRIEND_REQUEST_FAILURE",
  ACCEPT_FRIEND_REQUEST_START = "ACCEPT_FRIEND_REQUEST_START",
  ACCEPT_FRIEND_REQUEST_SUCCESS = "ACCEPT_FRIEND_REQUEST_SUCCESS",
  ACCEPT_FRIEND_REQUEST_FAILURE = "ACCEPT_FRIEND_REQUEST_FAILURE",
  REJECT_FRIEND_REQUEST_START = "REJECT_FRIEND_REQUEST_START",
  REJECT_FRIEND_REQUEST_SUCCESS = "REJECT_FRIEND_REQUEST_SUCCESS",
  REJECT_FRIEND_REQUEST_FAILURE = "REJECT_FRIEND_REQUEST_FAILURE",
  CANCEL_FRIEND_REQUEST_START = "CANCEL_FRIEND_REQUEST_START",
  CANCEL_FRIEND_REQUEST_SUCCESS = "CANCEL_FRIEND_REQUEST_SUCCESS",
  CANCEL_FRIEND_REQUEST_FAILURE = "CANCEL_FRIEND_REQUEST_FAILURE",
  REMOVE_FRIEND_START = "REMOVE_FRIEND_START",
  REMOVE_FRIEND_SUCCESS = "REMOVE_FRIEND_SUCCESS",
  REMOVE_FRIEND_FAILURE = "REMOVE_FRIEND_FAILURE",
  FETCH_RECEIVED_FRIEND_REQUESTS_START = "FETCH_RECEIVED_FRIEND_REQUESTS_START",
  FETCH_RECEIVED_FRIEND_REQUESTS_SUCCESS = "FETCH_RECEIVED_FRIEND_REQUESTS_SUCCESS",
  FETCH_RECEIVED_FRIEND_REQUESTS_FAILURE = "FETCH_RECEIVED_FRIEND_REQUESTS_FAILURE",
  RESET_FRIEND_REQUESTS_DATA = "RESET_FRIEND_REQUESTS_DATA",
}